:bug: Single-View constraints are now possible
When you previously tried to create a constraint on a view that related to itself (like: make the view’s height equal to the view's width) the constraint was actually installed on the superview. While this worked fine, it required the view to already be part of some view hierarchy, although a constraint like above could be treated like a characteristic of the view itself, unrelated from the view hierarchy the view might or might not end up in. The constraint is now installed on the view itself.